// src/app/shared/enums/malls-enum.ts
/* eslint-disable @typescript-eslint/no-unused-vars */
export enum MallsEnum {
  PAL = "PAL",
  PAN = "PAN",
  PCA = "PCA",
  PCO = "PCO",
  PEG = "PEG",
  PIQ = "PIQ",
  PLD = "PLD",
  PNO = "PNO",
  PTR = "PTR",
  PVE = "PVE",
}

export const enum MallsDescriptionEnum {
  alameda = "Mallplaza Alameda",
  antofagasta = "MallPlaza Antofagasta",
  calama = "Mallplaza Calama",
  copiapo = "Mallplaza Copiapó",
  dominicos = "Mallplaza Los Dominicos",
  egana = "Mallplaza Egaña",
  iquique = "Mallplaza Iquique",
  norte = "Mallplaza Norte",
  trebol = "Mallplaza Trébol",
  vespucio = "Mallplaza Vespucio",
}